refactor(monitoring): split performance-monitor into sub-modules (#594)
Decompose the 1198-line PerformanceMonitor god node into focused modules
(CPU, alerts, search stats, DB metrics, analytics) while preserving the
public API and passing all 26 monitoring regression tests.
